Updated blur percentage and tilt percentage
